python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audio_dsp"
version = "1.1.0"
description = "An audio DSP library for the xcore"
authors = [
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy >= 2.0",
    "scipy",
    "matplotlib < 3.9.1",
    "IPyWidgets",
    "graphviz",
    "IPython",
    "ipympl",
    "pyyaml",
    "mako",
    "tabulate",
    "docstring-inheritance",
    "packaging>=24.1",
    "pydantic"
]

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["D"]
ignore = [
    "D404", # First word of docstring must not be "This".
    "D205", # Requires blank line after first line of a doc string (specified by PEP 257).
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
